backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
prophet==1.1.6
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.1
supabase==2.11.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
joblib==1.4.2
